MDEV-18949: Analyze show ref(const,const) when we are using index scan
The issue here is that when we change the plan from REF access to index scan which can resolve the
ordering, then we were not resetting the ref keyparts and not updating the new estimates for the rows
to be read taking LIMIT into consideration.

This is happening for ANALYZE because we reset the ref keyparts to 0 only when the query is an EXPLAIN query.